+++ website/project/views/drafts.py
@@ -12,7 +12,7 @@ import pytz
 from modularodm import Q
 from modularodm.exceptions import ValidationValueError
 
-from framework.mongo.utils import get_or_http_error, autoload
+from framework.database import get_or_http_error, autoload
 from framework.exceptions import HTTPError
 from framework.status import push_status_message
